Bound hardware back button to navigating back. Separated files slightly.
