+++ src/App.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Provider } from 'react-redux'
-import { NavigatorComponent } from 'src/routes'
+import { NavigatorContainer } from 'src/NavigatorContainer'
 import { store } from 'src/store'
 
 const { ThemeProvider } = require ('react-native-material-ui')
@@ -11,7 +11,7 @@ export default class App extends React.Component {
     return (
       <ThemeProvider uiTheme={{}}>
         <Provider store={ store }>
-          <NavigatorComponent />
+          <NavigatorContainer />
         </Provider>
       </ThemeProvider>
     )
